+++ src/components/manage/BlockChooser/BlockChooser.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { filter, map, groupBy, isEmpty } from 'lodash';
 import { Accordion, Button } from 'semantic-ui-react';
-import { injectIntl, defineMessages } from 'react-intl';
+import { useIntl, defineMessages } from 'react-intl';
 import { Icon } from '@plone/volto/components';
 import AnimateHeight from 'react-animate-height';
 import config from '@plone/volto/registry';
@@ -28,9 +28,9 @@ const BlockChooser = ({
   allowedBlocks,
   showRestricted,
   blocksConfig = config.blocks.blocksConfig,
-  intl,
   blockChooserRef,
 }) => {
+  const intl = useIntl();
   const useAllowedBlocks = !isEmpty(allowedBlocks);
 
   const filteredBlocksConfig = filter(blocksConfig, (item) => {
@@ -150,9 +150,6 @@ BlockChooser.propTypes = {
   blocksConfig: PropTypes.objectOf(PropTypes.any),
 };
 
-export default injectIntl(
-  React.forwardRef((props, ref) => (
-    <BlockChooser {...props} blockChooserRef={ref} />
-  )),
-  { forwardRef: true },
-);
+export default React.forwardRef((props, ref) => (
+  <BlockChooser {...props} blockChooserRef={ref} />
+));